Add logfile to cached results
