--- tests/data/cache2.json
@@ -1,1 +1,0 @@
-{"mod_time": 1689532265.2332983, " -m HKY -redo": {"CandidateSet": {"0": "-6419.17009576 (0:0.0199914776,1:0.0227049948,2:0.3069481228);"}, "ModelDNA": {"rates": "1, 3.781004814, 1, 1, 3.781004814, 1"}, "PhyloTree": {"newick": "(0:0.0199914776,1:0.0227049948,2:0.3069481228);"}, "StopRule": {"curIteration": 101, "start_real_time": 180311.6276, "time_vec": null}, "boot_consense_logl": 0, "contree_rfdist": -1, "finished": true, "finishedCandidateSet": true, "finishedModelFinal": true, "finishedModelInit": true, "initTree": "(0:0.0238907850,1:0.0207730077,2:0.3489499192);", "iqtree": {"command": "-s /iqtree2/tests/tmpqanipmkt/data/three-ungapped.fa -m HKY -redo", "seed": 564928, "start_time": 1689817844, "version": "2.2.2.7"}}}
