+++ tests/test_fixtures.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 
-from .utils import iqtree1_dir, iqtree2_dir, repo_root, tests_data, tests_root
+from .test_utils import iqtree1_dir, iqtree2_dir, repo_root, tests_data, tests_root
 
 
 @pytest.mark.parametrize(
