+++ tests/test_utils/__init__.py
@@ -1,0 +1,7 @@
+from .checkpoint_file import Checkpoint
+from .execute import exec_command
+from .iqtree import Iqtree
+from .iqtree1 import Iqtree1
+from .iqtree2 import Iqtree2
+from .paths import iqtree1_dir, iqtree2_dir, repo_root, tests_data, tests_root
+from .log_file import Log
